@@ -6,26 +6,20 @@
   "author": "dry-tortuga",
   "license": "MIT",
   "dependencies": {
-    "@headlessui/react": "^1.4.3",
-    "@heroicons/react": "^1.0.5",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
-    "@web3-react/coinbase-wallet": "^8.2.3",
-    "@web3-react/core": "^8.2.3",
-    "@web3-react/metamask": "^8.2.4",
-    "@web3-react/walletconnect-v2": "^8.5.1",
+    "@tanstack/react-query": "^5.37.1",
     "bootstrap": "^5.1.3",
     "bootswatch": "^5.1.3",
     "buffer": "^6.0.3",
     "dayjs": "^1.11.9",
     "ethers": "^5.7.2",
-    "react": "^17.0.2",
+    "react": "^18.3.1",
     "react-bootstrap": "^2.4.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.3.1",
     "react-router-dom": "^6.2.2",
-    "react-scripts": "5.0.0",
+    "react-scripts": "^5.0.1",
     "react-select": "^5.7.2",
+    "viem": "^2.12.0",
+    "wagmi": "^2.9.6",
     "web3": "^1.7.1"
   },
   "devDependencies": {
